Check for token settings too
